GET request for questions
