+++ src/WebSocketContext.js
@@ -1,4 +1,3 @@
-// WebSocketContext.js
 import React, { createContext, useContext, useEffect, useState } from "react";
 import SockJS from "sockjs-client";
 import { over } from "stompjs";
@@ -6,6 +5,7 @@ import { over } from "stompjs";
 const WebSocketContext = createContext(null);
 
 export const useWebSocketContext = () => useContext(WebSocketContext);
+
 export const WebSocketProvider = ({ children }) => {
   const [stompClient, setStompClient] = useState(null);
   const BACKEND_URL = process.env.REACT_APP_BACKEND_URL;
@@ -24,7 +24,6 @@ export const WebSocketProvider = ({ children }) => {
 
   const onError = (err) => {
     console.error("WebSocket Error:", err);
-    // Try to reconnect after a delay
     setTimeout(connect, 3000); // Reconnect after 3 seconds
   };
